Refactor: refactor folder structure
