+++ components/Inputs/SelectInput/SelectInput.tsx
@@ -10,7 +10,7 @@ import {
 import { ErrorMessage } from '@hookform/error-message';
 import { CircleAlert } from 'lucide-react';
 import { useFormContext } from 'react-hook-form';
-import { Label } from '../ui/label';
+import { Label } from '../../ui/label';
 
 export type SelectOptions = {
   value: string;
